crud lancamento criado

+++ src/main/java/com/financas/api/payload/response/JwtResponse.java
@@ -6,12 +6,12 @@ public class JwtResponse {
 
   private String token;
   private String type = "Bearer";
-  private Long id;
+  private Integer id;
   private String username;
   private String email;
   private List<String> roles;
 
-  public JwtResponse(String accessToken, Long id, String username, String email, List<String> roles) {
+  public JwtResponse(String accessToken, Integer id, String username, String email, List<String> roles) {
     this.token = accessToken;
     this.id = id;
     this.username = username;
@@ -35,11 +35,11 @@ public class JwtResponse {
     this.type = tokenType;
   }
 
-  public Long getId() {
+  public Integer getId() {
     return id;
   }
 
-  public void setId(Long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
